requirements.txt
fastapi
uvicorn
sqlalchemy
loguru
passlib
bcrypt
python-multipart
aiofiles
python-jose
